# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "The Shoes Shop API",
    "version": "1.0.0",
    "description": "API for managing shoes inventory with JWT authentication and search capabilities"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Welcome endpoint",
        "responses": {
          "200": {
            "description": "Welcome message",
            "content": {
              "application/json": {
                "example": "Welcome to \"The Shoes Shop\" API."
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "john_doe"
                  },
                  "password": {
                    "type": "string",
                    "example": "securepassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "example": { "message": "User registered successfully" }
              }
            }
          },
          "400": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "User already exists"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Authenticate user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "john_doe"
                  },
                  "password": {
                    "type": "string",
                    "example": "securepassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthToken"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid username or password"
                }
              }
            }
          }
        }
      }
    },
    "/shoes": {
      "get": {
        "tags": ["Shoes"],
        "summary": "Get all shoes",
        "responses": {
          "200": {
            "description": "List of shoes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shoe"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Shoes"],
        "summary": "Create a new shoe",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShoeInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shoe created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shoe"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/search-shoes": {
      "get": {
        "tags": ["Shoes"],
        "summary": "Search shoes by query parameters",
        "parameters": [
          {
            "name": "brand",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "color",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "price",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching shoes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shoe"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/shoes/{id}": {
      "get": {
        "tags": ["Shoes"],
        "summary": "Get shoe by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shoe details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shoe"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": ["Shoes"],
        "summary": "Update shoe details",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShoeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated shoe details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shoe"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["Shoes"],
        "summary": "Delete a shoe",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Shoe deleted successfully"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Shoe": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "abc123"
          },
          "brand": {
            "type": "string",
            "example": "Nike"
          },
          "model": {
            "type": "string",
            "example": "Air Max"
          },
          "size": {
            "type": "number",
            "example": 10.5
          },
          "color": {
            "type": "string",
            "example": "Black"
          },
          "price": {
            "type": "number",
            "example": 129.99
          }
        }
      },
      "ShoeInput": {
        "type": "object",
        "required": ["brand", "model", "size", "color", "price"],
        "properties": {
          "brand": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "color": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "AuthToken": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid token."
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "noToken": {
                "value": { "error": "Access denied. No token provided." }
              },
              "invalidToken": {
                "value": { "error": "Invalid token." }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Shoe not found"
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Authentication",
      "description": "User registration and authentication"
    },
    {
      "name": "Shoes",
      "description": "Shoes inventory management"
    }
  ]
}